requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-google-genai==0.0.11
pinecone-client==3.0.0
langchain-pinecone==0.1.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.2
python-dotenv==1.0.0
google-generativeai==0.3.1
python-multipart==0.0.6
